requirements.txt
opencv-contrib-python==4.10.0.84
scipy==1.14.1
numpy==1.23.5
pyttsx3==2.98
mysql-connector-python==9.1.0
imutils==0.5.4
deap==1.4.1
